docs + more refactor
